+++ CHANGELOG.rst
@@ -2,3 +2,4 @@ Unreleased
 ----------
 * new method: `Introduction.introduction`
 * new method: `Introduction.ombud_messages`
+* fixed: `database.add_instrument_to_db`